+++ picamtracker/MotionTracker.py
@@ -49,6 +49,8 @@ import cv2
 from math import atan2,hypot,degrees,acos,pi,sqrt
 from time import sleep,time
 import prctl
+import enum
+import logging
 
 #- globals
 MAX_TRACKS     = 16
@@ -88,6 +90,13 @@ def normalize_angle(x):
     if x > np.pi:
         x -= 2 * np.pi
     return(x)
+
+class direction(enum.Enum):
+    unknown = 0
+    left2right = 1
+    right2left = 2
+    up2down = 3
+    down2up = 4
 
 class Tracker(threading.Thread):
     """
@@ -182,7 +191,7 @@ class Tracker(threading.Thread):
             ym = int(self.rows / 2)
             #         [target_rect    ]         [bounding_rect          ]
             motion = [[xm-5,ym-5,10,10],[10,10],[xm-10,ym-10,xm+10,ym+10]]
-            
+
             frame = self.camera.frame.index
             self.crossed(99,time(),frame,motion)
 
@@ -207,6 +216,7 @@ class Tracker(threading.Thread):
             self.frame = 0
             self.direction = 0
             self.detectionDelay = -0.999
+        #print("Tracker::getStatus (%d)" % frame)
         return (delay, frame, motion)
 
     #--------------------------------------------------------------------
@@ -219,14 +229,14 @@ class Tracker(threading.Thread):
 
         with self.lock:
             self.locked = True
-            self.camera.request_key_frame()
             if self.udpThread:
                 self.udpThread.event.set()
             if self.greenLEDThread:
                 self.greenLEDThread.event.set()
+            self.detectionDelay = time() - timestamp
+            self.camera.request_key_frame()
             self.updates  = updates
             self.frame  = frame
-            self.detectionDelay = time() - timestamp
             self.motion = motion
             self.direction = 1 if positive_direction else -1
 
@@ -428,6 +438,7 @@ class Track:
         self.distance = [0,0]
         self.dirxOK = False
         self.diryOK = False
+        self.direction = 0
         self.distxOK = False
         self.distyOK = False
         self.old_dir  = None
@@ -473,6 +484,26 @@ class Track:
                 #print "[%s] (%d) clean" % (self.name, self.updates)
                 self.reset()
 
+    def leadingEdge(self,rn):
+
+        # right to left -> left edge is leading
+        if self.direction == 1:
+             return rn[0], rn[1] + rn[3]/2
+        # left to right -> right edge is leading
+        if self.direction == 2:
+             return rn[0] + rn[2], rn[1] + rn[3]/2
+
+        # up to down -> lower edge is leading
+        if self.direction == 4:
+             return rn[0] + rn[2]/2, rn[1] + rn[3]
+        # down to up -> upper edge is leading
+        if self.direction == 3:
+             return rn[0] + rn[2]/2, rn[1]
+
+        # unknown
+        return rn[0] + rn[2]/2, rn[1] + rn[3]/2
+
+
     def new_track(self,timestamp,frame,rn,vn):
         """
         start a new track
@@ -485,8 +516,29 @@ class Track:
         #    #print("[%s]: %d/%d %3.1f/%3.1f new track rejected (too high)" % (self.name, rn[0]+rn[2],rn[1]+rn[3],vn[0],vn[1]))
         #    return 0
 
-        cxn  = rn[0]+rn[2]/2 #xn+wn/2
-        cyn  = rn[1]+rn[3]/2 #yn+hn/2
+        # determine followup type
+        # xCross > 0 and x > xmax / 2 -> right_to_left
+        # xCross > 0 and x < xmax / 2 -> left_to_right
+        # yCross > 0 and y > ymax / 2 -> up_to_down
+        # yCross > 0 and y < ymax / 2 -> down_to_up
+
+        if Track.xCross > 0:
+            if rn[0] > int(Track.maxX / 2):
+                self.direction = 1
+            else:
+                self.direction = 2
+
+        if Track.yCross > 0:
+            if rn[1] > int(Track.maxY / 2):
+                self.direction = 3
+            else:
+                self.direction = 4
+
+
+        cxn,cyn = self.leadingEdge(rn)
+
+        #cxn  = rn[0]+rn[2]/2 #xn+wn/2
+        #cyn  = rn[1]+rn[3]/2 #yn+hn/2
 
         #cxn  = rn[0]
         #cyn  = rn[1]
@@ -508,27 +560,29 @@ class Track:
         """
 
         if self.updates >= 3:
-
+            delta = Track.maxDist
             if Track.xCross > 0:
                 x0 = int(self.tr[0][0])
                 x1 = int(self.tr[2][0])
-                xc = int(Track.maxX / 2)
+                #xc = int(Track.maxX / 2)
                 #- if the first occurence is right from center movement must be negative
-                if x0 >= xc and (x1-x0) <= 0:
+                #- objects occuring too near to the center are rejected
+                if x0 > (Track.xCross + delta) and x1 < x0:
                     return True
                 #- if the first occurence is left from center movement must be positive
-                if x0 < xc and (x1-x0) >= 0:
+                if x0 < (Track.xCross - delta) and x1 >= x0:
                     return True
 
             if Track.yCross > 0:
                 y0 = int(self.tr[0][1])
                 y1 = int(self.tr[2][1])
-                yc = int(Track.maxY / 2)
+                #yc = int(Track.maxY / 2)
                 #- if the first occurence is above center movement must be negative
-                if y0 >= yc and (y1-y0) <= 0:
+                #- objects occuring too near to the center are rejected
+                if y0 > (Track.yCross + delta) and y1 < y0:
                     return True
                 #- if the first occurence is below center movement must be positive
-                if y0 < yc and (y1-y0) >= 0:
+                if y0 < (Track.yCross - delta) and y1 > y0:
                     return True
 
                 #print("[%s](%d) start failed: yc: %d dy: %d" % (self.name, self.updates, y0, (y1-y0)))
@@ -637,9 +691,12 @@ class Track:
     #-- TODO: make the same for x direction
     #--------------------------------------------------------------------
     def detectCrossing(self, dx, dy, r):
-        delta = 3
+        delta = 3 # validate this!
+        min_coverage = 0.4 # validate this!
+        min_fill_grade = 0.25 # validate this!
+        speed_limit = 5.0 # validate this!
 
-        if Track.yCross > 0:
+        if Track.yCross > 0 and self.crossedY == False and self.progressy:
             # track is crossing target line in Y direction
             #  x0,y0 +--------------+
             #        |              |
@@ -648,40 +705,114 @@ class Track:
             #  v > 0 |          --->|
             #        |              |
             #  v < 0 |<---          |
-            #print("   [%s](%02d) x0/y0:%d/%d x1/y1:%d/%d vy:%3.1f/%3.1f dy:%d/%d maturity: %d"
-            #        % (self.name,self.updates,r[0],r[1],r[0]+r[2],r[1]+r[3],-self.vv[0],-self.vv[1],dy,dx, self.maturity))
-            #print("   %4.2f > 0.4?" % (self.distance[1] / self.deltaY))
-            if self.updates >= self.maturity and self.progressy == True and self.crossedY == False:
-                # develope indicators
-                vx = -self.vv[0]; vy = -self.vv[1] # remember the velocity has wrong direction!
-                x0 = r[0]; y0 = r[1]
-                y1 = r[1] + r[3]
+            
+            # variables
+            vx = -self.vv[0]; vy = -self.vv[1] # remember the velocity has wrong direction!
+            x0 = r[0]; y0 = r[1]
+            y1 = r[1] + r[3]
+            # for low speeds take distance as indicator
+            vy_ = abs(vy)
+            if vy_ < 0.1:
+                vy = float(dy)
+            if vy_ > delta:
+                delta = int(vy_) + 1
+            # moving quality
+            coverage = self.distance[1] / self.deltaY
 
-                # for low speeds take distance as indicator
-                vy_ = abs(vy)
-                if vy_ < 0.1:
-                    vy = float(dy)
-                if vy_ > 5:
-                    delta = int(vy_/2) + 1
-
+            # mature check
+            crossedYPositive = crossedYNegative = False
+            fastText = ''
+            if self.updates >= self.maturity:
                 # this model uses a simple >= limit to detect a crossing event
-                crossedYPositive =  vy >  0.1 and y1 >= Track.yCross and (y1 - delta) < Track.yCross and self.miny < Track.yCross and self.deltaY > self.maturity #and self.distance[1] / self.deltaY > 0.4
-                crossedYNegative =  vy < -0.1 and y0 <= Track.yCross and (y0 + delta) > Track.yCross and self.maxy > Track.yCross and self.deltaY > self.maturity #and -self.distance[1] / self.deltaY > 0.4
+                crossedYPositive =  vy >  0.1 and y1 >= Track.yCross and (y1 - delta) < Track.yCross and self.miny < Track.yCross and coverage > min_coverage
+                crossedYNegative =  vy < -0.1 and y0 <= Track.yCross and (y0 + delta) > Track.yCross and self.maxy > Track.yCross and coverage < -min_coverage
+            else:
+                # fast crossing check for big and fast objects
+                # a) object is faster than speed limit
+                # b) object is longer than half of the distance side to turn
+                # c) object front is over the turn line
+                # d) object end is behind the turn line 
+                fill_grade = r[3] / Track.maxY / 2
+                if fill_grade > min_fill_grade and y0 <= Track.yCross and y1 >= Track.yCross:
+                    crossedYPositive = vy > speed_limit
+                    crossedYNegative = vy < -speed_limit
+                    if crossedYPositive or crossedYNegative:
+                        fastText = 'FAST-'
+                        
+            if crossedYPositive:
+                delay = (time() - self.timestamp) * 1000.0
+                
+                print("[%s](%02d/%4.1f) y1:%2d/%2d vy:%+5.1f/%+5.1f dy:%2d/%2d deltaY:%2d dist:%3d cov:%4.1f %sY-CROSSED++++++++++++++++++++"
+                    % (self.name,self.updates,delay,y1,x0,vy,vx,dy,dx,self.deltaY,self.distance[1],coverage,fastText))
+                self.crossedY = True
+                self.crossed(positive=True)
 
-                if crossedYPositive:
-                    delay = (time() - self.timestamp) * 1000.0
-                    print("[%s](%02d/%3.0f) y1:%d/%d vy:%3.1f/%3.1f dy:%d/%d delta:%d dist:%d Y-CROSSED++++++++++++++++++++"
-                        % (self.name,self.updates,delay,y1,x0,vy,vx,dy,dx,self.deltaY,self.distance[1]))
-                    self.crossedY = True
-                    self.crossed(positive=True)
+            if crossedYNegative:
+                delay = (time() - self.timestamp) * 1000.0
+                print("[%s](%02d/%4.1f) y0:%2d/%2d vy:%+5.1f/%+5.1f dy:%2d/%2d deltaY:%2d dist:%3d cov:%4.1f %sY-CROSSED--------------------"
+                    % (self.name,self.updates,delay,y0,x0,vy,vx,dy,dx,self.deltaY,self.distance[1],coverage,fastText))
+                self.crossedY = True
+                self.crossed(positive=False)
 
-                if crossedYNegative:
-                    delay = (time() - self.timestamp) * 1000.0
-                    print("[%s](%02d/%3.0f) y0:%d/%d vy:%3.1f/%3.1f dy:%d/%d delta:%d dist:%d Y-CROSSED--------------------"
-                        % (self.name,self.updates,delay,y0,x0,vy,vx,dy,dx,self.deltaY,self.distance[1]))
-                    self.crossedY = True
-                    self.crossed(positive=False)
+        if Track.xCross > 0 and self.crossedX == False and self.progressx:
+            # track is crossing target line in Y direction
+            #  x0,y0 +--------------+
+            #        |              |
+            #        +--------------+ x1,y1
+            #
+            #  v > 0 |          --->|
+            #        |              |
+            #  v < 0 |<---          |
+            
+            # variables
+            vx = -self.vv[0]; vy = -self.vv[1] # remember the velocity has wrong direction!
+            x0 = r[0]; y0 = r[1]
+            x1 = r[0] + r[2]
+            # for low speeds take distance as indicator
+            vx_ = abs(vx)
+            if vx_ < 0.1:
+                vx = float(dx)
+            if vx_ > delta:
+                delta = int(vx_) + 1
+            # moving quality
+            coverage = self.distance[0] / self.deltaX
 
+            # mature check
+            crossedXPositive = crossedXNegative = False
+            fastText = ''
+            if self.updates >= self.maturity:
+                # this model uses a simple >= limit to detect a crossing event
+                crossedXPositive =  vx >  0.1 and x1 >= Track.xCross and (x1 - delta) < Track.xCross and self.minx < Track.xCross and coverage > min_coverage
+                crossedXNegative =  vx < -0.1 and x0 <= Track.xCross and (x0 + delta) > Track.xCross and self.maxx > Track.xCross and coverage < -min_coverage
+            else:
+                # fast crossing check for big and fast objects
+                # a) object is faster than speed limit
+                # b) object is longer than half of the distance side to turn
+                # c) object front is over the turn line
+                # d) object end is behind the turn line 
+                fill_grade = r[2] / Track.maxX / 2
+                if fill_grade > min_fill_grade and x0 <= Track.xCross and x1 >= Track.xCross:
+                    crossedXPositive = vx > speed_limit
+                    crossedXNegative = vx < -speed_limit
+                    if crossedXPositive or crossedXNegative:
+                        fastText = 'FAST-'
+                        
+            if crossedXPositive:
+                delay = (time() - self.timestamp) * 1000.0
+                
+                print("[%s](%02d/%4.1f) x1:%2d/%2d vx:%+5.1f/%+5.1f dx:%2d/%2d deltaX:%2d dist:%3d cov:%4.1f %sX-CROSSED++++++++++++++++++++"
+                    % (self.name,self.updates,delay,x1,y0,vx,vy,dx,dy,self.deltaX,self.distance[0],coverage,fastText))
+                self.crossedX = True
+                self.crossed(positive=True)
+
+            if crossedXNegative:
+                delay = (time() - self.timestamp) * 1000.0
+                print("[%s](%02d/%4.1f) x0:%2d/%2d vx:%+5.1f/%+5.1f dx:%2d/%2d deltaX:%2d dist:%3d cov:%4.1f %sY-CROSSED--------------------"
+                    % (self.name,self.updates,delay,x0,y0,vx,vy,dx,dy,self.deltaX,self.distance[0],coverage,fastText))
+                self.crossedY = True
+                self.crossed(positive=False)
+
+        """
         if Track.xCross > 0:
             # track is crossing target line in X direction
             #  x0,y0 +--------------+
@@ -702,27 +833,31 @@ class Track:
                 vx_ = abs(vx)
                 if vx_ < 0.1:
                     vx = float(dx)
-                if vx_ > 5:
-                    delta = int(vx_/2) + 1
+                if vx_ > delta:
+                    delta = int(vx_) + 1
+
+                # moving quality
+                coverage = self.distance[0] / self.deltaX
 
                 # this model uses a simple >= limit to detect a crossing event
-                crossedXPositive =  vx >  0.1 and x1 >= Track.xCross and (x1 - delta) < Track.xCross and self.minx < Track.xCross and self.deltaX > self.maturity
-                crossedXNegative =  vx < -0.1 and x0 <= Track.xCross and (x0 + delta) > Track.xCross and self.maxx > Track.xCross and self.deltaX > self.maturity
+                crossedXPositive =  vx >  0.1 and x1 >= Track.xCross and (x1 - delta) < Track.xCross and self.minx < Track.xCross and coverage > 0.4
+                crossedXNegative =  vx < -0.1 and x0 <= Track.xCross and (x0 + delta) > Track.xCross and self.maxx > Track.xCross and coverage < -0.4
 
                 if crossedXPositive:
                     delay = (time() - self.timestamp) * 1000.0
-                    print("[%s](%02d/%3.0f) x1:%d/%d vx:%3.1f/%3.1f dx:%d/%d X-CROSSED++++++++++++++++++++"
+                    print("[%s](%02d/%4.1f) x1:%d/%d vx:%3.1f/%3.1f dx:%d/%d X-CROSSED++++++++++++++++++++"
                         % (self.name,self.updates,delay,x1,y0,vx,vy,dx,dy))
                     self.crossedX = True
                     self.crossed(positive=True)
 
                 if crossedXNegative:
                     delay = (time() - self.timestamp) * 1000.0
-                    print("[%s](%02d/%3.0f) x0:%d/%d vx:%3.1f/%3.1f dx:%d/%d X-CROSSED--------------------"
+                    print("[%s](%02d/%4.1f) x0:%d/%d vx:%3.1f/%3.1f dx:%d/%d X-CROSSED--------------------"
                         % (self.name,self.updates,delay,x0,y0,vx,vy,dx,dy))
                     self.crossedX = True
                     self.crossed(positive=False)
-
+        """
+        
     #--------------------------------------------------------------------
     #-- does the track leave the tracking area?
     #--------------------------------------------------------------------
@@ -749,11 +884,10 @@ class Track:
         self.timestamp = timestamp
 
         # PiMotionAnalysis.analyse may be called more than once per frame -> double hit?
-        cxn  = rn[0]+rn[2]/2.0
-        cyn  = rn[1]+rn[3]/2.0
+        #cxn  = rn[0]+rn[2]/2.0
+        #cyn  = rn[1]+rn[3]/2.0
 
-        # estimating via the object center produces very much noise. (??? I didn't think much about that)
-        # estimation is done via the upper left corner (until I have something better)
+        cxn,cyn = self.leadingEdge(rn)
 
         #cxn  = rn[0]
         #cyn  = rn[1]
@@ -765,30 +899,27 @@ class Track:
             #print "[%s] double hit" % self.name
             return self.id
 
+
+
+        # reject objects which change area too much from frame to frame
+        area  = rn[2] * rn[3]
+        if self.old_area is None:
+            self.old_area = area
+        if area > self.old_area:
+            delta_area = area/self.old_area
+        else:
+            delta_area = self.old_area/area
+        if delta_area > 15.0 and self.updates > 2:
+            return 0x00000000
+        self.old_area = area
+        found = 0x00000000
+
         # try to append new coordinates to track
-        # TODO: use vx/vy to improve accuracy
         dx    = cxn - self.cx
         dy    = cyn - self.cy
         vx    = -vn[0]
         vy    = -vn[1]
         dist  = hypot(dx,dy)
-        area  = rn[2] * rn[3]
-
-        if self.old_area is None:
-            self.old_area = area
-
-        # reject objects which change area too much from frame to frame
-        if area > self.old_area:
-            delta_area = area/self.old_area
-        else:
-            delta_area = self.old_area/area
-
-        # reject too big area changes
-        if delta_area > 15.0:
-            return 0x00000000
-
-        self.old_area = area
-        found = 0x00000000
 
         # poor mans perspective
         # big nearby objects may move fast --- far away objects may move slow
@@ -796,7 +927,9 @@ class Track:
         max_dist = Track.maxDist
         fill_grade = area / Track.maxArea
 
-        if(fill_grade > 0.2):
+        # fast crossing check
+        if(fill_grade > 0.3 and self.updates > 2):
+            speed = hypot(vx,vy)
             m_factor = 10. if (fill_grade > 0.9) else 1.0 / (1.0 - fill_grade)
             #- enhace distance for big objects (just an empiric estimation)
             max_dist *= int(2.0 * m_factor)
@@ -838,12 +971,7 @@ class Track:
             else:
                 cos_delta = 1.0
 
-            # >>> debug
-            #print("%s,%5d,%2d,%2d,%2d,%2d,%2d,%4.2f,%4.2f" %
-            #      (self.name,frame,self.updates,rn[0],rn[1],rn[2],rn[3],cos_delta,dist))
-            # <<< debug
-            # reject all tracks out of direction
-            #if self.updates < 3 or dist <= 1.1 or abs(cos_delta) > Track.minCosDelta:
+            # check direction. (the first three tracks may have a wron direction)
             if self.updates < 3 or abs(cos_delta) > Track.minCosDelta:
                 #- update base data
                 found = self.id
@@ -855,14 +983,14 @@ class Track:
                 self.cy = cyn
                 self.updates += 1
                 self.tr.append([cxn,cyn])
-                if(len(self.tr) > 24):
+                if(len(self.tr) > 36):
                     del self.tr[0]
 
                 #- stop investigation if object is moving in the wrong direction
                 if self.updates == 3:
                     if not self.checkStartingConditions():
                         self.reset()
-                        return self.id
+                        return 0x00000000
 
                 #- is the coverered area still growing?
                 self.updateGrowingStatus(rn)
@@ -870,11 +998,11 @@ class Track:
                 #- does the track leave the playground?
                 if self.isLeaving(dx,dy):
                     self.reset()
-                    return self.id
-
+                    return 0x00000000
 
                 # crossing status
-                self.detectCrossing(dx,dy,rn)
+                if self.updates >= 2:
+                    self.detectCrossing(dx,dy,rn)
 
                 # turning status (disabled for now)
                 # self.detectTurn(dx,dy,rn)